Update execution tests (#243)
* Updated failing execution tests: semanticTests/various.

* Updated failing execution tests: semanticTests/viaYul.

* Updated failing execution tests: semanticTests/viaYul (cont'd).

* Updated failing compilation tests.

+++ test/libsolidity/semanticTests/various/nested_calldata_struct_to_memory.sol
@@ -3,20 +3,20 @@ pragma abicoder               v2;
 
 contract C {
     struct S1 {
-        uint256 a;
-        uint256 b;
+        uint a;
+        uint b;
     }
     struct S2 {
-        uint256 a;
-        uint256 b;
+        uint a;
+        uint b;
         S1 s;
-        uint256 c;
+        uint c;
     }
 
     function f(S2 calldata s)
         external
         pure
-        returns (uint256 a, uint256 b, uint256 sa, uint256 sb, uint256 c)
+        returns (uint a, uint b, uint sa, uint sb, uint c)
     {
         S2 memory m = s;
         return (m.a, m.b, m.s.a, m.s.b, m.c);
@@ -26,4 +26,4 @@ contract C {
 // ====
 // compileViaYul: also
 // ----
-// f((uint256,uint256,(uint256,uint256),uint256)): 1, 2, 3, 4, 5 -> 1, 2, 3, 4, 5
+// f((uint,uint,(uint,uint),uint)): refargs { 1, 2, 3, 4, 5 } -> 1, 2, 3, 4, 5
